updated: センサロック完了（ただし，コミットする前にポインタ式を廃止，動確は未検証）          並進系，角速度系制御の実装中          sinf を使用できない，角速度系の加算未対応

+++ UserSrc/FunctionTimer.h
@@ -14,7 +14,7 @@
 void FTR_Init(void);
 
 void FTR_StartBuzzerTimer(void);
-void FTR_SetBuzzerDuty(uint16_t duty);
+void FTR_SetBuzzerDuty(float32_t duty);
 
 void FTR_StartRightEncoderTimer(void);
 uint16_t FTR_GetRightEncoderCount(void);
@@ -22,15 +22,15 @@ void FTR_StartLeftEncoderTimer(void);
 uint16_t FTR_GetLeftEncoderCount(void);
 
 void FTR_StartRightMotorTimer(void);
-void FTR_SetRightMotorDuty(uint16_t duty);
+void FTR_SetRightMotorDuty(float32_t duty);
 uint16_t FTR_GetRightMotorDuty(void);
 
 void FTR_StartLeftMotorTimer(void);
-void FTR_SetLeftMotorDuty(uint16_t duty);
+void FTR_SetLeftMotorDuty(float32_t duty);
 uint16_t FTR_GetLeftMotorDuty(void);
 
 void FTR_StartSensorMotorTimer(void);
-void FTR_SetSensorMotorDuty(uint16_t duty);
+void FTR_SetSensorMotorDuty(float32_t duty);
 uint16_t FTR_GetSensorMotorDuty(void);
 
 #endif /* FUNCTIONTIMER_H_ */
